pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "vector-etl"
authors = [
    {name = "Jide Ogunjobi", email = "jide@contextdata.ai"},
]
description = "Lightweight ETL Framework for Vector Databases"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["vector database", "etl"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
   "boto3",
   "botocore",
   "cohere",
   "cffi",
   "openai",
   "psycopg2-binary",
   "pinecone",
   "requests",
   "tiktoken",
   "python-dotenv",
   "pydantic",
   "mysql-connector-python",
   "pymysql",
   "pandas",
   "qdrant-client",
   "singlestoredb",
   "weaviate-client",
   "azure-storage-blob",
   "google-cloud-storage",
   "snowflake-connector-python",
   "stripe",
   "vecs",
   "simple-salesforce",
   "anthropic",
   "pympler",
   "unstructured[all-docs]",
   "dropbox",
   "zenpy",
   "lancedb",
   "pyyaml",
   "google-auth",
   "google-auth-oauthlib",
   "google-auth-httplib2",
   "google-api-python-client",
   "unstructured-client",
   "box-sdk-gen",
   "pymongo",
   "neo4j",
   "python-magic",
   "pytest",
   "nltk",
   "pymilvus",
   "pygithub>=2.8.1",
   "google-genai>=1.66.0",
   "chromadb>=1.5.2",
   "langchain-text-splitters>=1.1.1",
   "numpy>=2.4.2",
]
dynamic = ["version"]

[project.scripts]
vector-etl = "vector_etl.main:main"

[tool.setuptools_scm]
